+++ gas-optimization/erc721-staking/test/implementation/OptimizedStaking721.sol
@@ -27,7 +27,7 @@ contract OptimizedStaking721Impl is IStaking721Mock, OptimizedStaking721, IERC72
         revert("Unsupported function");
     }
 
-    function setStakingCondition(uint256 _timeUnit, uint256 _rewardsPerUnitTime) external {
+    function setStakingCondition(uint32 _timeUnit, uint128 _rewardsPerUnitTime) external {
         _setStakingCondition(_timeUnit, _rewardsPerUnitTime);
     }
