+++ gas-optimization/erc721-staking/test/interface/IStaking721Mock.sol
@@ -4,5 +4,5 @@ pragma solidity 0.8.21;
 import {IStaking721} from "../../src/interface/IStaking721.sol";
 
 interface IStaking721Mock is IStaking721 {
-    function setStakingCondition(uint256 _timeUnit, uint256 _rewardsPerUnitTime) external;
+    function setStakingCondition(uint32 _timeUnit, uint128 _rewardsPerUnitTime) external;
 }
